rework architecture
