+++ Cargo.toml
@@ -1,36 +1,33 @@
 [package]
 name = "embedbot-rs"
-version = "0.5.0"
+version = "0.6.0"
 authors = ["Liss Heidrich"]
 resolver = "2"
 edition = "2021"
 
 [dependencies]
-anyhow = "1.0.86"
-thiserror = "1.0.61"
-clap = { version = "4.5.4", features = ["derive", "string", "cargo"] }
-tracing = "0.1.40"
-tracing-subscriber = "0.3.18"
-reqwest = { version = "0.12.4", default-features = false, features = ["rustls-tls", "json"] }
-serde = { version = "1.0.203", features = ["derive"] }
-serde_json = "1.0.117"
-tokio = { version = "1.38.0", features = ["macros", "rt-multi-thread", "signal"] }
-url = "2.5.0"
-serenity = { version = "0.12.2", default-features = false, features = ["rustls_backend", "model", "client", "gateway"] }
+anyhow = "1.0.95"
+thiserror = "2.0.9"
+clap = { version = "4.5.23", features = ["derive", "string", "cargo"] }
+tracing = "0.1.41"
+tracing-subscriber = "0.3.19"
+reqwest = { version = "0.12.9", default-features = false, features = ["rustls-tls", "json"] }
+serde = { version = "1.0.216", features = ["derive"] }
+serde_json = "1.0.134"
+toml = "0.8.19"
+tokio = { version = "1.42.0", features = ["macros", "rt-multi-thread", "signal"] }
+url = "2.5.4"
+serenity = { version = "0.12.4", default-features = false, features = ["rustls_backend", "model", "client", "gateway"] }
 itertools = "0.13.0"
 v_escape = "0.18.0"
 
-resvg = { version = "0.40.0", optional = true }
-tempfile = { version = "3.10.1", optional = true }
-
-scraper = { version = "0.19.0", optional = true }
+scraper = { version = "0.22.0", optional = true }
 json_nav = { git = "https://github.com/Clueliss/json_nav", optional = true }
 headless_chrome = { version = "1.0.10", optional = true }
 
 
 [features]
-default = ["reddit", "ninegag", "svg", "twitter", "implicit-auto-embed"]
-svg = ["dep:resvg", "dep:tempfile"]
+default = ["reddit", "ninegag", "twitter", "implicit-auto-embed"]
 reddit = ["dep:json_nav"]
 ninegag = ["dep:scraper", "dep:json_nav"]
 twitter = ["dep:scraper", "dep:headless_chrome"]
