--- embedbot.json
@@ -1,32 +1,0 @@
-{
-    "discord_token": "YOUR_DISCORD_TOKEN",
-    "modules": {
-        "reddit": {
-            "embed_set": [
-                {
-                    "content_type": "Image"
-                },
-                {
-                    "content_type": "Gallery"
-                },
-                {
-                    "content_type": "Text"
-                },
-                {
-                    "content_type": "Video"
-                }
-            ]
-        },
-        "ninegag": {
-            "embed_set": [
-                "Image"
-            ]
-        },
-        "svg": {},
-        "twitter": {
-            "embed_set": [
-                "Video"
-            ]
-        }
-    }
-}
